test(infra): run warp ALT override tests in CI
